refactor: move util.GetMountPathOfPod to pod resources package
